+++ .github/workflows/e2e-test-releases-reusable-pipeline.yml
@@ -95,6 +95,8 @@ on:
         required: true
       BOOTSTRAP_DEV_PROXY:
         required: true
+      DEV_MODULES_REGISTRY_PASSWORD:
+        required: true
 
 env:
   BRANCH: ${{ inputs.branch }}
@@ -1385,6 +1387,95 @@ jobs:
           kubectl -n d8-virtualization wait --for=condition=Available deploy --all --timeout 900s
           echo "[INFO] Checking virt-handler pods"
           virt_handler_ready
+
+      - name: Setup crane and registry auth
+        uses: deckhouse/modules-actions/setup@v2
+        with:
+          registry: ${{ vars.DEV_REGISTRY }}
+          registry_login: ${{ vars.DEV_MODULES_REGISTRY_LOGIN }}
+          registry_password: ${{ secrets.DEV_MODULES_REGISTRY_PASSWORD }}
+
+      - name: Verify image digests in pods after upgrade
+        run: |
+          MODULE_IMAGE="${{ vars.DEV_REGISTRY }}/${{ vars.DEV_MODULE_SOURCE }}/virtualization:${{ env.NEW_RELEASE }}"
+          echo "[INFO] Extracting images_digests.json from ${MODULE_IMAGE}"
+          images_hash=$(crane export "${MODULE_IMAGE}" - | tar -Oxf - images_digests.json)
+          echo "[INFO] Expected image digests:"
+          echo "::group::images_digests.json"
+          echo "$images_hash" | jq .
+          echo "::endgroup::"
+
+          audit_status=$(kubectl get mc virtualization -o=jsonpath='{.spec.settings.audit.enabled}' 2>/dev/null || true)
+          audit_image_skip="true"
+          if [ -n "$audit_status" ] && [ "$audit_status" == "true" ]; then
+            audit_image_skip="false"
+          fi
+
+          SKIP_IMAGES=()
+          if [ "$audit_image_skip" == "true" ]; then
+            SKIP_IMAGES+=("virtualizationAudit")
+          fi
+          SKIP_IMAGES+=("virtualizationDraUsb")
+
+          is_skipped_image() {
+            local img="$1"
+            if [ ${#img} -eq 0 ]; then return 1; fi
+            for skip in "${SKIP_IMAGES[@]}"; do
+              if [[ "$img" == "$skip" ]]; then
+                return 0
+              fi
+            done
+            return 1
+          }
+
+          retry_count=0
+          max_retries=120
+          sleep_interval=5
+
+          while true; do
+            all_hashes_found=true
+
+            v12n_pods=$(kubectl -n d8-virtualization get pods -o json | jq -c)
+
+            while IFS= read -r image_entry; do
+              image=$(echo "$image_entry" | jq -r '.key')
+              hash=$(echo "$image_entry" | jq -r '.value')
+
+              if [[ "${image,,}" =~ (libguestfs|predeletehook) ]]; then
+                continue
+              fi
+
+              if is_skipped_image "$image"; then
+                echo "- SKIP $image"
+                continue
+              fi
+
+              if echo "$v12n_pods" | grep -q "$hash"; then
+                echo "- OK   $image $hash"
+              else
+                echo "- MISS $image $hash"
+                all_hashes_found=false
+              fi
+            done < <(echo "$images_hash" | jq -c '. | to_entries | sort_by(.key)[]')
+
+            if [ "$all_hashes_found" = true ]; then
+              echo "[SUCCESS] All image hashes found in pods after upgrade to ${{ env.NEW_RELEASE }}"
+              break
+            fi
+
+            retry_count=$((retry_count + 1))
+            echo "[INFO] Some hashes are missing, rechecking... Attempt: ${retry_count}/${max_retries}"
+
+            if [ "$retry_count" -ge "$max_retries" ]; then
+              echo "[ERROR] Timeout reached after $((retry_count * sleep_interval))s. Some image hashes are still missing."
+              echo "::group::pods in d8-virtualization"
+              kubectl -n d8-virtualization get pods -o wide || true
+              echo "::endgroup::"
+              exit 1
+            fi
+
+            sleep "$sleep_interval"
+          done
 
       - name: Show MPO state after upgrade
         run: |
